Turned form into a component.
