+++ src/components/atoms.jsx
@@ -1,4 +1,14 @@
 import { atom } from "jotai";
+import axios from "axios";
+
+const fetchWarehouseData = async () => {
+  try {
+    const response = await axios.get("/api/warehouse");
+    return response.data;
+  } catch (error) {
+    throw error;
+  }
+};
 
 export const accessTokenAtom = atom(null);
 export const refreshTokenAtom = atom(null);
@@ -12,4 +22,15 @@ export const isAuthAtom = atom(
 export const removeTokensAtom = atom(null, (get, set, updatedAccessToken) => {
   set(accessTokenAtom, null);
   set(refreshTokenAtom, null);
+});
+
+export const warehouseDataAtom = atom([]);
+
+export const updateWarehouseDataAtom = atom(null, async (get, set, updated) => {
+  try {
+    const data = await fetchWarehouseData();
+    set(warehouseDataAtom, data);
+  } catch (error) {
+    console.error(error);
+  }
 });
